requirements.txt
streamlit
azure-storage-blob
snowflake-connector-python

